system string to wchar_t* conversion for wrapper

+++ minispy_kernel/minispy.c
@@ -1139,7 +1139,7 @@ Return Value:
                     break;
                 }
 
-                if (ClusterMapBuffer.Extents[0].NextVcn.QuadPart - ClusterMapBuffer.StartingVcn.QuadPart < 0ll) {
+                if (ClusterMapBuffer.Extents[0].NextVcn.QuadPart - ClusterMapBuffer.StartingVcn.QuadPart < (UINT64)0) {
                     DbgPrint("Region was lower than 0, reg len :%X, extents %X, StartingVcn : %X, name : %wZ\n", ClusterMapBuffer.Extents[0].NextVcn.QuadPart - ClusterMapBuffer.StartingVcn.QuadPart, ClusterMapBuffer.Extents[0].NextVcn.QuadPart, ClusterMapBuffer.StartingVcn.QuadPart, &nameInfo->Name);
                     
                     for (int i = 0; i < RecCount; i++) {
